requirements.txt
EdgeGPT-fork
pyTelegramBotAPI==4.14.0